requirements.txt
requests
transformers
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
